[#3225] Split WithEntitlements into WithPlanEntitlements and WithMaterializedLimits
